Move modal into VocabDetailscomponent and set up pages folder for form and list page
