+++ package.json
@@ -16,6 +16,7 @@
     "react-router-dom": "5.1.2",
     "react-scripts": "3.4.0",
     "react-test-renderer": "16.13.0",
+    "react-uid": "2.2.0",
     "styled-components": "5.0.1"
   },
   "scripts": {
