+++ src/App.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import styled from 'styled-components/macro'
-import Vocab from './components/Vocab/Vocab'
-import data from './vocabs.json'
 import Modal from 'react-modal'
+import { BrowserRouter as Router, Route, Switch, Link } from 'react-router-dom'
+import FormPage from './pages/FormPage'
+import ListPage from './pages/ListPage'
+import data from './vocabs.json'
 import PrimaryButton from './components/PrimaryButton/PrimaryButton'
-import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom'
-import VocabForm from './components/VocabForm/VocabForm'
+import styled from 'styled-components/macro'
 
 Modal.setAppElement(document.getElementById('root'))
 
@@ -13,34 +13,23 @@ export default function App() {
   const vocabs = data.vocabs
 
   return (
-    <Router>
-      <AppGrid>
+    <AppGrid>
+      <Router>
         <Switch>
           <Route exact path="/">
-            <VocabList>
-              {vocabs.map(vocab => (
-                <Vocab
-                  imageSrc={vocab.imageSrc}
-                  audioSrc={vocab.audioSrc}
-                  wordTitle={vocab.wordTitle}
-                  translation={vocab.translation}
-                  partOfSpeechCategory={vocab.partOfSpeechCategory}
-                  key={vocab.id}
-                />
-              ))}
-            </VocabList>
-            <Navigation>
-              <Link to="/create">
-                <PrimaryButton buttonLabel="Add new Vocabulary" />
-              </Link>
-            </Navigation>
+            <ListPage vocabs={vocabs} />
           </Route>
           <Route path="/create">
-            <VocabForm />
+            <FormPage />
           </Route>
         </Switch>
-      </AppGrid>
-    </Router>
+        <Navigation>
+          <Link to="/create">
+            <PrimaryButton buttonLabel="Add new Vocabulary" />
+          </Link>
+        </Navigation>
+      </Router>
+    </AppGrid>
   )
 }
 
@@ -48,22 +37,6 @@ const AppGrid = styled.div`
   display: grid;
   grid-template-rows: 1fr 8fr 1fr;
   height: 100vh;
-`
-
-const VocabList = styled.section`
-  grid-row-start: 2;
-  grid-row-end: 3;
-  display: flex;
-  flex-wrap: nowrap;
-  overflow: scroll;
-  padding: 0 30px 30px;
-  scroll-snap-type: x mandatory;
-  /* scroll-padding: 0 300px 0 0; */
-
-  :after {
-    content: '';
-    flex: 0 0 10px;
-  }
 `
 
 const Navigation = styled.section`
